Add CalendarView and DateView
